requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
alpaca-py